All done, waiting for my boys to finish their job

+++ api/src/routes/checkout.js
@@ -11,7 +11,7 @@ server.post('/:userId', (req, res) => {
     const itemsToMP = products.map(item => {
         return {
             title: item.name,
-            unit_price: parseInt(item.price - (item.price * (item.discount / 100))),
+            unit_price: parseFloat(item.price - (item.price * (item.discount / 100))),
             quantity: item.order_line.quantity
         };
     });
